backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
sqlalchemy>=2.0.27,<3.0.0
chromadb>=0.4.24,<0.6.0
google-genai>=0.2.0
crewai==0.28.8
langchain>=0.1.10,<0.2
langchain-google-genai>=0.0.6
python-dotenv>=1.0.0
pyPDF2
